requirements_mvp.txt
feedparser>=6.0.10
beautifulsoup4>=4.12.2
langdetect>=1.0.9
python-telegram-bot>=20.7
openai>=1.3.0
requests>=2.31.0
python-dateutil>=2.8.2
lxml>=5.0.0
pytz>=2023.3
